biology function test refactor
