+++ janitor/testing_utils/fixtures.py
@@ -1,6 +1,10 @@
 import numpy as np
 import pandas as pd
 import pytest
+from pathlib import Path
+
+
+TEST_DATA_DIR = Path(__file__).resolve().parent.parent.parent / "tests" / "test_data"
 
 
 @pytest.fixture
@@ -73,4 +77,10 @@ def missingdata_df():
         "cities": ["Cambridge", "Shanghai", "Basel"] * 3,
     }
     df = pd.DataFrame(data)
+    return df
+
+
+@pytest.fixture
+def biodf():
+    df = pd.read_csv(TEST_DATA_DIR / "sequences.tsv", sep="\t").clean_names()
     return df
