+++ tests/biology/test_join_fasta.py
@@ -1,0 +1,13 @@
+import janitor.biology
+import pytest
+
+from janitor.testing_utils.fixtures import TEST_DATA_DIR, biodf
+
+
+@pytest.mark.biology
+def test_join_fasta(biodf):
+    df = biodf.join_fasta(
+        TEST_DATA_DIR / "sequences.fasta", "sequence_accession", "sequence"
+    )
+
+    assert "sequence" in df.columns
